# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Image Codex",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/compositions/": {
      "post": {
        "tags": [
          "compositions"
        ],
        "summary": "Create Composition",
        "description": "Create an image composition and return it in the given type",
        "operationId": "create_composition_compositions__post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Mimetype",
              "type": "string",
              "default": "application/pdf"
            },
            "name": "mimetype",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Composition"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Get Images",
        "description": "Get images with filters:\n- **tags**: contains all given tags\n- **author**: has given author",
        "operationId": "get_images_images__get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Tags",
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": []
            },
            "name": "tags",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Author",
              "type": "string"
            },
            "name": "author",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Next",
              "type": "string"
            },
            "name": "next",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "title": "Size",
              "maximum": 100.0,
              "minimum": 1.0,
              "type": "integer",
              "default": 50
            },
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_TaggedImage_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Create Image",
        "description": "Upload a new image to the database",
        "operationId": "create_image_images__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiFile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/images/{image_ids}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "Get Image",
        "description": "Get images with given ids",
        "operationId": "get_image_images__image_ids__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Image Ids",
              "type": "string"
            },
            "name": "image_ids",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Image Images  Image Ids  Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaggedImage"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tags/": {
      "get": {
        "tags": [
          "tags"
        ],
        "summary": "Get Tags",
        "description": "Get all tags",
        "operationId": "get_tags_tags__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Tags Tags  Get",
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiFile": {
        "title": "ApiFile",
        "required": [
          "base64"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "base64": {
            "title": "Base64",
            "type": "string"
          }
        }
      },
      "ComposedImage": {
        "title": "ComposedImage",
        "required": [
          "id",
          "url",
          "x",
          "y",
          "width",
          "height"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          }
        }
      },
      "Composition": {
        "title": "Composition",
        "required": [
          "name",
          "width",
          "height",
          "images"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "background_color": {
            "title": "Background Color",
            "type": "array",
            "items": {
              "type": "integer"
            },
            "default": [
              255,
              255,
              255
            ]
          },
          "images": {
            "title": "Images",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComposedImage"
            }
          }
        }
      },
      "CursorPage_TaggedImage_": {
        "title": "CursorPage[TaggedImage]",
        "required": [
          "items",
          "total",
          "size"
        ],
        "type": "object",
        "properties": {
          "items": {
            "title": "Items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaggedImage"
            }
          },
          "total": {
            "title": "Total",
            "minimum": 0.0,
            "type": "integer"
          },
          "next": {
            "title": "Next",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "minimum": 1.0,
            "type": "integer"
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "TaggedImage": {
        "title": "TaggedImage",
        "required": [
          "id",
          "name",
          "url",
          "width",
          "height",
          "tags",
          "author",
          "license"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "tags": {
            "title": "Tags",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "author": {
            "title": "Author",
            "type": "string"
          },
          "license": {
            "title": "License",
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}